+++ drift_dev/test/writer/schema_version_writer_test.dart
@@ -6,6 +6,8 @@ import 'package:drift_dev/src/writer/schema_version_writer.dart';
 import 'package:drift_dev/src/writer/writer.dart';
 import 'package:test/test.dart';
 
+import '../analysis/test_utils.dart';
+
 void main() {
   final fakeUri = Uri.parse('drift:hidden');
 
@@ -98,5 +100,69 @@ void main() {
     for (final tableName in secondaryProblemTables.map((t) => t.baseDartName)) {
       expect(output, matches(containsTableRegex(tableName, withSuffix: true)));
     }
+  });
+
+  test('can generate references between columns', () async {
+    // Regression test for https://github.com/simolus3/drift/issues/3554
+    final backend = await TestBackend.inTest({
+      'a|lib/a.dart': r'''
+import 'package:drift/drift.dart';
+import 'package:drift/extensions/json1.dart';
+
+class ComplexTable extends Table {
+  @override
+  String get tableName => 'complex_table';
+
+  // A json column
+  TextColumn get reference => text()();
+
+ // extracts data from the reference column
+  TextColumn get packageId =>
+      text()
+          .generatedAs(reference.jsonExtract<String>(r'$.package_id'), stored: true)();
+}
+''',
+    });
+
+    final results = await backend.analyze('package:a/a.dart');
+    backend.expectNoErrors();
+
+    final [table as DriftTable] = results.analyzedElements.toList();
+
+    final imports = LibraryImportManager();
+    final writer = Writer(
+      const DriftOptions.defaults(),
+      generationOptions: GenerationOptions(imports: imports),
+    );
+    imports.linkToWriter(writer);
+    SchemaVersionWriter(
+      [
+        SchemaVersion(
+          1,
+          [table],
+          const {},
+        ),
+        SchemaVersion(
+          2,
+          [table],
+          const {},
+        ),
+      ],
+      writer.child(),
+    ).write();
+
+    final output = writer.writeGenerated();
+
+    // Writing the reference in jsonExtract should not write `reference`
+    // directly, because that's not possible when extracting columns from the
+    // table.
+    expect(
+      output,
+      contains(
+        "generatedAs: i1.GeneratedAs(i2.JsonExtensions("
+        "(i0.VersionedTable.col<String>('reference'))"
+        ").jsonExtract<String>(r'\$.package_id')",
+      ),
+    );
   });
 }